requirements.txt
psycopg2==2.9.6
fastapi==0.100.0
uvicorn==0.23.1
passlib==1.7.4
PyJWT==2.8.0
pytest==7.4.0
httpx==0.24.1
requests==2.31.0